backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]==3.2.9
django-cors-headers==4.3.1
python-decouple==3.8 